fix(codegen,build): Linux cross-platform support, parallel move resolution, GNU-stack
- Fix coalescer dependency check direction in PX_COPY resolution — was
  checking if source is another task's destination instead of checking
  if destination is another task's source, causing write-before-read
  hazards in entry parameter shuffles (fixes crash in multi-param
  entity init calls)
- Replace entry parameter MOV loop with PX_COPY pseudo-instruction
  to leverage topological sort and cycle-breaking for correct parallel
  move resolution (LowerILToMIR.cpp)
- Make label counter module-wide instead of per-function to prevent
  duplicate assembler labels (fixes vtris build)
- Add .note.GNU-stack section to ELF assembly output for non-executable
  stack marking (Backend.cpp)
- Add -lX11 linking on Linux for graphics demos (LinkerSupport.cpp)
- Guard macOS frameworks behind #if defined(__APPLE__) in AArch64
  backend (cmd_codegen_arm64.cpp)
- Make build_demos.sh cross-platform (arm64/x64 detection, bash
  arithmetic fix)
- Add Linux build support: CMake conditional targets, missing headers,
  POSIX macros, GUI stub implementations, graphics platform stubs
- Update codegen tests to accept peephole-eliminated XMM identity moves

+++ src/frontends/basic/Parser.hpp
@@ -60,6 +60,7 @@
 #include "support/diag_expected.hpp"
 #include "support/source_manager.hpp"
 #include <array>
+#include <cstdint>
 #include <functional>
 #include <initializer_list>
 #include <memory>
@@ -149,7 +150,7 @@ class Parser
 
     // Tracked constants discovered during parsing (for SELECT CASE label resolution).
     // Keys are canonicalised via CanonicalizeIdent for case-insensitive lookup.
-    std::unordered_map<std::string, long long> knownConstInts_{};
+    std::unordered_map<std::string, int64_t> knownConstInts_{};
     std::unordered_map<std::string, std::string> knownConstStrs_{};
 
     struct NamedLabelEntry
